Rework stylesheets to have posts "sit" correctly
Modified skeleton stylesheet and main stylesheet so the auto-generating
posts sat correctly on the page when displayed. Originally there was
some kind of margin affecting the way they sat (the next row was off by
about 10px or so) so after some painstaking trial and error they
finally sit as I want them to (for now)!